pyproject.toml
[tool.poetry]
name = "fundvis"
version = "0.1.0"
description = ""
authors = ["phil harper <phil@imrge.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.32.0"
networkx = "^3.2.1"
matplotlib = "^3.8.3"
supabase = "^2.4.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.28"
psycopg2 = "^2.9.9"
watchdog = "^4.0.0"
ipykernel = "^6.29.3"
plotly = "^5.19.0"
scipy = "^1.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
